pyproject.toml
[project]
name = "scientific-writer"
version = "2.2.2"
description = "AI-powered scientific writing with programmatic API and CLI - powered by Claude Sonnet 4.5 and the Claude Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

# System Dependencies (must be installed separately):
# 
# LaTeX Distribution (required for PDF compilation):
#   - macOS: MacTeX (https://www.tug.org/mactex/) or BasicTeX
#     Install: brew install --cask mactex-no-gui
#   - Linux: TeX Live
#     Install: sudo apt-get install texlive-full (Ubuntu/Debian)
#              sudo dnf install texlive-scheme-full (Fedora)
#   - Windows: MiKTeX (https://miktex.org/) or TeX Live
#
# Required LaTeX packages (usually included in full distributions):
#   - pdflatex, bibtex, latexmk
#   - Common packages: amsmath, graphicx, hyperref, natbib, geometry
#
# Optional but recommended:
#   - Ghostscript (for PDF manipulation)
#   - ImageMagick (for image processing)

[project.scripts]
scientific-writer = "scientific_writer.cli:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scientific_writer"]

[tool.uv]
dev-dependencies = []

